upgrade class structures
