+++ libsql/syntax/keywords.py
@@ -10,10 +10,28 @@ from typing import Union
 class KeywordABC:
     """ Keyword abstract class """
 
+    @classmethod
+    def _make(cls, val):
+        if isinstance(val, cls):
+            return val
+        if isinstance(val, str):
+            return cls(val.encode().upper())
+        if isinstance(val, bytes):
+            return cls(val.upper())
+        raise TypeError('Invalid type %s (%s)' % (type(val), val))
+
 
 class OrderType(KeywordABC, Enum):
     ASC = b'ASC'
     DESC = b'DESC'
+
+    @classmethod
+    def make(cls, val) -> 'OrderType':
+        if val is True:
+            return OrderType.ASC
+        if val is False:
+            return OrderType.DESC
+        return super()._make(val)
 
 
 class JoinType(KeywordABC, Enum):
@@ -23,6 +41,10 @@ class JoinType(KeywordABC, Enum):
     OUTER = b'OUTER'
     CROSS = b'CROSS'
 
+    @classmethod
+    def make(cls, val) -> 'JoinType':
+        return super()._make(val)
+
 
 class ReferenceOption(KeywordABC, Enum):
     RESTRICT = b'RESTRICT'
@@ -30,30 +52,10 @@ class ReferenceOption(KeywordABC, Enum):
     SET_NULL = b'SET_NULL'
     NO_ACTION = b'NO_ACTION'
 
+    @classmethod
+    def make(cls, val) -> 'ReferenceOption':
+        return super()._make(val)
 
 OrderLike = Union[OrderType, bool, bytes, str]
-
-def make_order_type(val: OrderLike):
-    if isinstance(val, OrderType):
-        return val
-    if val is True:
-        return OrderType.ASC
-    if val is False:
-        return OrderType.DESC
-    if isinstance(val, str):
-        return OrderType(val.encode().upper())
-    if isinstance(val, bytes):
-        return OrderType(val.upper())
-    raise TypeError('Invalid type %s (%s)' % (type(val), val))
-
-
 JoinLike = Union[JoinType, bytes, str]
-
-def make_join_type(val: JoinLike):
-    if isinstance(val, JoinType):
-        return val
-    if isinstance(val, str):
-        return JoinType(val.encode().upper())
-    if isinstance(val, bytes):
-        return JoinType(val.upper())
-    raise TypeError('Invalid type %s (%s)' % (type(val), val))
+RefOptionLike = Union[ReferenceOption, bytes, str]
